refactor(setup): simplify useCompleteAgentSetup and UnfinishedSetupAlert
Address review feedback from Tanya on PR #1839:

- Replace hasAttemptedCreation ref with isLoadingMasterSafeCreation (isPending)
  guard; the ref toggle dance in handleTryAgain was fragile.
- Inline the needsFunding navigation inside the hook and drop the
  shouldNavigateToFundYourAgent / resetShouldNavigate command-channel state;
  UnfinishedSetupAlert is now purely presentational.
- Remove redundant `= {}` default on AgentSetupCompleteModal's destructured
  optional props.
- Update test selector for the Complete Setup loading state to query by
  accessible name instead of the brittle `.ant-btn-loading` class.

No user-visible behavior change.
